\name{se.report}
\alias{se.report}
\title{Report Generation for Side Effect Predictions}
\usage{
se.report(results, file = "Rplots.pdf", ...)
}
\arguments{
\item{results}{List of predicted side effects}
\item{file}{Output filename}
\item{...}{Additional arguments passed to 'pdf()'}
}
\description{
'se.report' generates scatterplots correlating gene expressions and side effects based on expression profiles created by pipoft's other side effect-related functions.}
\keyword{side effects, predict, correlations}

